successful fix with tests seemingly passing and explained
